Move VersionInfo into the includes directory
